requirements.txt
torch
torchvision
timm
opencv-python
albumentations
numpy
pillow
matplotlib
onnx
onnxruntime
gradio
streamlit
fastapi
uvicorn[standard]
pydantic
scikit-learn
umap-learn
scikit-image
